requirements.txt
Flask==3.0.0
flask-cors==4.0.0
python-docx==1.1.0
fpdf==1.7.2
requests==2.31.0
python-dotenv==1.0.0
typing==3.7.4.3
PyPDF2==3.0.1  # For PDF text extraction
python-multipart==0.0.9